+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 8.8.4 (2025-08-01)
+
+[22f8dc3]: https://github.com/erqk/ng-dynamic-json-form/commit/22f8dc3e2bea791c92d540e6d4da80d3a2a2d07e
+[c91e975]: https://github.com/erqk/ng-dynamic-json-form/commit/c91e975386a155453d2c2022b7c241fd4be7aee1
+
+| Commit    | Type | Description                                                                              |
+| --------- | ---- | ---------------------------------------------------------------------------------------- |
+| [22f8dc3] | fix  | FormControlComponent: Don't set errors for custom component on instantiate.              |
+| [c91e975] | feat | PropsBindingDirective: Check if property is signal, and bind only valid HTML attributes. |
+
 # 8.8.3 (2025-07-04)
 
 [e6b7c04]: https://github.com/erqk/ng-dynamic-json-form/commit/e6b7c049c36d21d010506e10e5840e09bb86d6f3